Add extra validation of qubes-dom0-update output
The salt module do parse some of the output when querying repository
(things like listing available packages, latest version etc). In all the
cases we use this module, that information is later validated against
actual system state in dom0 (comparing if given package was installed
etc), but it is possible to call the salt module directly to just
retrieve the information. Add validation if package metadata matches
allowed charset for each relevant field.