Abort latest_version() if downloading pkg list fails
Previously it aborted only if some package was missing, but failed to
set 'updates' variable. That was partial change in
QubesOS/qubes-issues#6585 (before which, it aborted only on non-empty
stderr). Finish the change to abort on any metadata download failure. It
may result in failures that in some cases would be harmless, but
aborting on any metadata download failure makes it less likely to
falsely report no updates if there are some.

Fixes QubesOS/qubes-issues#7962
Related to QubesOS/qubes-issues#6585